backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
python-dotenv
pydantic
cors
beautifulsoup4
httpx
pinecone-client
langchain-huggingface
sentence-transformers
playwright
langchain-pinecone
pymupdf
pdfplumber
